+++ Questionable/Utils/LocalizeShortcut.cs
@@ -51,6 +51,7 @@ internal static class LocalizeShortcut
             CabinetSubCategory cabinetSubCategory => cabinetSubCategory.Name.ToMacroString(),
             FittingShopCategory fittingShopCategory => fittingShopCategory.Name.ToMacroString(),
             ContentRoulette contentRoulette => contentRoulette.Name.ToMacroString(),
+            HowTo howTo => howTo.Name.ToMacroString(),
             _ => throw new InvalidOperationException($"No known Name/Text mapping for {typeof(T).Name}")
         };
         _translatedStrings[(typeof(T), rowId)] = value;